frame/inspect.rs: extract pure header-parsing leaf fns (#988 Phase 2a)
Move 10 read-only header-parsing fns from frame/mod.rs into a new
sibling frame/inspect.rs:

- frame_l3_offset / frame_l4_offset
- packet_rel_l4_offset / packet_rel_l4_offset_and_protocol
- extract_tcp_flags_and_window / extract_tcp_window
- frame_has_tcp_rst
- tcp_flags_str
- metadata_tuple_complete
- parse_flow_ports

All ten are pure (no side effects, no internal state). Visibility
matches the prior re-export width: frame_has_tcp_rst stays
pub(in crate::afxdp) because afxdp.rs / tx/transmit.rs /
cos/queue_service.rs reach for it; the others are pub(super) /
pub(in crate::afxdp) per where they're called from.

frame/mod.rs production LOC (excluding mod tests):
  2,804 → 2,487 (−317)
inspect.rs: 349 LOC.

The larger inspect cluster (parse_session_flow_*,
authoritative_forward_ports, try_parse_metadata) stays in mod.rs
for now and lands in Phase 2b. Per docs/refactor/988-frame-decomp.md
(origin/research/988-frame-design). Phase 0/1 (checksum.rs) in
#1019/#1021.

build clean (cargo build --release), 865 Rust tests pass.
